+++ Commands/Enums/ConnectionType.cs
@@ -2,7 +2,6 @@
 {
     public enum ConnectionType
     {
-        OnPrem = 0,
         O365 = 1,
         TenantAdmin = 2,
     }